Add imagemagick
This change might be a bit controversial.

When using the Gitlab Pipelines we have to possibilities. One is very
easy: Just pick the ruby base image and run the tests in there. There
other one is to build your own image and to the whole setup with
docker-compose.

This change enables knowledge-base base to be built with the first
solution. It makes the whole process easier. But it adds another
dependency to the base image which is not needed for most other apps.

Fat image vs. easier build process. Pick your poison.